fix: move some types to dedicated files
