+++ src/index.ts
@@ -2,4 +2,4 @@
  * @public
  */
 export { SDJwtPlugin } from './agent-plugin/sd-jwt-plugin.js';
-export * from './types/i-sd-jwt-plugin.js';
+export * from './types/ISDJwtPlugin.js';
